+++ crate/object_model/src/config/object/character/sequence_id.rs
@@ -1,8 +1,9 @@
 /// Object Sequence IDs.
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq)]
+#[serde(rename_all = "snake_case")]
 pub enum SequenceId {
     /// Default sequence for characters.
-    Standing,
+    Stand,
     /// Walking sequence.
-    Walking,
+    Walk,
 }