+++ crate/object_model/src/entity/character/mod.rs
@@ -1,0 +1,5 @@
+//! Contains the game play types for character entities.
+
+pub use self::character_input::CharacterInput;
+
+mod character_input;