+++ crate/object_play/src/character/mod.rs
@@ -1,0 +1,4 @@
+pub use self::character_sequence_handler::CharacterSequenceHandler;
+
+mod character_sequence_handler;
+mod sequence_handler;