+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # Changelog
 
+## 0.3.0 (master, unreleased)
+
+* Read controller configuration.
+* Update character sequence based on input.
+
 ## 0.2.0 (2018-04-26)
 
 * Loads object sprites and animations.
 * Creates an entity per object.
 
 ## 0.1.0 (2018-03-24)
-
-**Will:**
 
 * Loads fonts and displays a simple menu.
 * Published on [itch.io](https://azriel91.itch.io/will)